+++ tests/unit/entity/test_item.py
@@ -39,6 +39,7 @@ class TestItem(UnitTestCase):
         # THEN
         self.assertEqual(item.title, "Item Title")
         self.assertEqual(item.price, 1234.99)
+        self.assertEqual(item.asin, "B0018CJYCO")
 
     def test_title_starts_with_ampersand_use_lxml(self):
         # GIVEN
@@ -61,3 +62,22 @@ class TestItem(UnitTestCase):
 
         # THEN
         self.assertEqual(item.title, "&And Per Se Lined")
+        self.assertEqual(item.asin, "B0CW5Y6PKG")
+
+    def test_asin_none_when_link_not_a_product_page(self):
+        # GIVEN an item whose link is not a product detail page (e.g. a "Buy it again" URL)
+        html = """
+<div class="a-fixed-left-grid-col yohtmlc-item a-col-right">
+<div class="a-row">
+<a class="a-link-normal" href="/gp/buyagain?ie=UTF8&amp;ref_=ppx_od_dt_b_buy_again">Item Title</a>
+</div>
+</div>
+"""
+        parsed = BeautifulSoup(html, self.test_config.bs4_parser)
+
+        # WHEN
+        item = Item(parsed, self.test_config)
+
+        # THEN
+        self.assertEqual(item.title, "Item Title")
+        self.assertIsNone(item.asin)